# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "learnwithai",
    "description": "HTTP API for LearnWithAI, including operational endpoints, UNC authentication flows, and authenticated user profile access.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/courses/{course_id}/exam-pdfs": {
      "post": {
        "tags": [
          "Exam PDFs"
        ],
        "summary": "Upload an exam PDF",
        "description": "Validates and stores an uploaded exam PDF.\n\nArgs:\n    subject: Authenticated user uploading the file.\n    course: Course loaded via path dependency.\n    exam_pdf_svc: Service handling authorization and persistence.\n    file: Uploaded PDF from multipart form-data.\n\nReturns:\n    Stored upload metadata including storage key.\n\nRaises:\n    HTTPException: If validation fails or storage encounters an error.",
        "operationId": "upload_exam_pdf",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_exam_pdf"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The stored upload metadata and storage key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamPdfUploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid file."
          },
          "401": {
            "description": "Not authenticated."
          },
          "403": {
            "description": "Insufficient permissions."
          },
          "404": {
            "description": "Course not found."
          },
          "500": {
            "description": "Storage failure."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Exam PDFs"
        ],
        "summary": "List exam PDF uploads for the current student",
        "description": "Retrieves all exam PDF uploads for the requesting student in a course.\n\nArgs:\n    subject: Authenticated user requesting the list.\n    course: Course loaded via path dependency.\n    exam_pdf_upload_repo: Repository for loading upload records.\n    practice_material_repo: Repository for checking practice material existence.\n    membership_repo: Repository for checking course membership.\n\nReturns:\n    List of exam uploads for the student, newest first.\n\nRaises:\n    HTTPException: If user lacks permissions or course is not found.",
        "operationId": "list_exam_pdf_uploads",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All exam uploads by the requesting student in this course, newest first.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExamPdfHistoryItem"
                  },
                  "title": "Response List Exam Pdf Uploads"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated."
          },
          "403": {
            "description": "Insufficient permissions."
          },
          "404": {
            "description": "Course not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{course_id}/exam-pdfs/{upload_id}/analysis": {
      "get": {
        "tags": [
          "Exam PDFs"
        ],
        "summary": "Retrieve exam analysis",
        "description": "Retrieves the analysis results for an uploaded exam PDF.\n\nArgs:\n    subject: Authenticated user requesting the analysis.\n    course: Course loaded via path dependency.\n    upload_id: Upload identifier from the URL path.\n    exam_pdf_upload_repo: Repository for loading upload records.\n\nReturns:\n    Structured analysis with topics, strengths, and weaknesses.\n\nRaises:\n    HTTPException: If the upload is not found, analysis is not available, or user lacks permission.",
        "operationId": "get_exam_analysis",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "upload_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Upload Id"
            }
          },
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Structured analysis of an uploaded exam PDF.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamPdfAnalysisResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated."
          },
          "403": {
            "description": "Insufficient permissions."
          },
          "404": {
            "description": "Upload not found or analysis not available."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_exam_pdf": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_exam_pdf"
      },
...
      "ExamPdfAnalysisResponse": {
        "properties": {
          "upload_id": {
            "type": "integer",
            "title": "Upload Id"
          },
          "analysis_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Analysis Data"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "upload_id",
          "analysis_data",
          "created_at"
        ],
        "title": "ExamPdfAnalysisResponse",
        "description": "Response containing analysis results for an uploaded exam PDF."
      },
      "ExamPdfHistoryItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Uploaded At"
          },
          "has_analysis": {
            "type": "boolean",
            "title": "Has Analysis"
          },
          "has_practice": {
            "type": "boolean",
            "title": "Has Practice"
          }
        },
        "type": "object",
        "required": [
          "id",
          "original_filename",
          "uploaded_at",
          "has_analysis",
          "has_practice"
        ],
        "title": "ExamPdfHistoryItem",
        "description": "Response item for exam PDF history list."
      },
      "ExamPdfUploadResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "course_id": {
            "type": "integer",
            "title": "Course Id"
          },
          "uploader_pid": {
            "type": "integer",
            "title": "Uploader Pid"
          },
          "storage_key": {
            "type": "string",
            "title": "Storage Key"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_id",
          "uploader_pid",
          "storage_key",
          "original_filename",
          "content_type",
          "size_bytes",
          "created_at"
        ],
        "title": "ExamPdfUploadResponse",
        "description": "Response returned when a PDF upload is stored successfully."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}